requirements.txt
absl-py==1.4.0
albucore==0.0.23
albumentations==2.0.5
anaconda-client==1.12.0
anaconda-navigator==2.4.2
annotated-types==0.7.0
appdirs==1.4.4
apriori==1.0.0
astunparse==1.6.3
atomicwrites==1.4.0
attrs==20.3.0
backports.weakref==1.0.post1
blosc2==2.0.0
brotlipy==0.7.0
cachelib==0.12.0
cachetools==5.3.1
check50==3.3.11
clyent==1.2.2
cmake==3.31.4
coloredlogs==15.0.1
conda-repo-cli==1.0.41
conda-verify==3.4.2
constantly==15.1.0
crypto==1.4.1
cs50==9.3.4
cssselect==1.1.0
Cython==3.0.11
daal4py==2023.1.1
datashape==0.5.4
dlib==19.24.6
docker-pycreds==0.4.0
EditorConfig==0.12.3
et-xmlfile==1.1.0
Flask-Session==0.8.0
flatbuffers==24.3.25
fonttools==4.25.0
fsspec==2024.3.1
fst-pso==1.8.1
FuzzyTM==2.0.9
gast==0.4.0
gaze-tracking==0.0.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.58.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.22.2
humanfriendly==10.0
icdiff==2.0.7
inflection==0.5.1
jax==0.4.37
jaxlib==0.4.36
jsbeautifier==1.14.11
jsonpointer==2.1
jupyterthemes==0.20.0
kaggle==1.6.17
kagglehub==0.3.10
keras==3.7.0
lesscpy==0.15.1
lib50==3.0.8
libclang==16.0.6
lit==18.1.8
llvmlite==0.43.0
mediapipe==0.10.20
miniful==0.0.6
mkl-service==2.4.0
ml-dtypes==0.4.1
mlxtend==0.23.0
msgspec==0.18.6
munkres==1.1.4
mypy-extensions==0.4.3
Naked==0.1.32
namex==0.0.8
navigator-updater==0.4.0
ndindex==1.9.2
nltk==3.8.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnx==1.13.0
onnx-simplifier==0.4.36
onnx2keras==0.0.24
onnxconverter-common==1.14.0
onnxruntime==1.20.0
opencv-contrib-python==4.10.0.84
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
opendatasets==0.1.22
openpyxl==3.0.10
opt-einsum==3.3.0
optree==0.14.0
pandas==1.5.3
patsy==0.5.3
pdf2image==1.17.0
pep8==1.7.1
Pillow==9.4.0
ply==3.11
preprocessing==0.1.13
protobuf==3.20.3
py4j==0.10.9.7
pyarrow==11.0.0
pyasn1-modules==0.2.8
pycryptodome==3.21.0
pycurl==7.45.2
pydantic==2.10.6
pydantic_core==2.27.2
PyDispatcher==2.0.5
pyFUME==0.3.1
pygame==2.5.2
pyls-spyder==0.4.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
pyspark==3.5.2
pytesseract==0.3.13
python-lsp-jsonrpc==1.0.0
python-magic==0.4.27
queuelib==1.5.0
requests-oauthlib==1.3.1
rich==13.7.1
rsa==4.9
scikit-learn-intelex==20230426.111612
sentencepiece==0.2.0
sentry-sdk==2.0.1
setproctitle==1.3.3
shellescape==3.8.1
simpful==2.12.0
simsimd==6.2.1
sip==6.8.6
smmap==5.0.1
sounddevice==0.5.1
sphinx-rtd-theme==0.2.4
sqlparse==0.4.4
stringzilla==3.11.3
style50==2.9.1
submit50==3.1.2
sympy==1.13.1
TBB==0.2
tensorboard==2.18.0
tensorboard-data-server==0.7.1
tensorflow==2.18.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow_decision_forests==1.11.0
termcolor==1.1.0
tf2onnx==1.16.1
tf_keras==2.18.0
tk==0.1.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchinfo==1.8.0
torchvision==0.20.1+cu121
triton==3.1.0
ttkbootstrap==1.10.1
typing_extensions==4.12.2
wandb==0.16.6
webencodings==0.5.1
xgboost==2.1.1
ydf==0.9.0
